# docs/openapi/claim-service.v1.json
{
  "components" : {
    "parameters" : {
      "CorrelationId" : {
        "description" : "Optional. Traces the request across services and Kafka; generated if absent and always returned in the response header and error bodies.",
        "in" : "header",
        "name" : "X-Correlation-ID",
        "required" : false,
        "schema" : {
          "example" : "7f92a1c2-8e32-4e12-b7d2-91a3f8c4d111",
          "type" : "string"
        }
      }
    },
    "schemas" : {
      "AdjusterResponse" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiError" : {
        "properties" : {
          "correlationId" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "violations" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "CreateAdjusterRequest" : {
        "properties" : {
          "email" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "email", "name" ],
        "type" : "object"
      },
      "FieldViolation" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "First Notice of Loss, the claim lifecycle, adjusters and the audit trail (ClaimCenter-like). Supports Idempotency-Key on FNOL and ETag/If-Match on changes.\n\nAPI version **v1** (contract 1.0.0). Breaking changes get a new `/api/v2`; additions keep v1 and bump the minor version.",
    "license" : {
      "name" : "Learning project, not for production use"
    },
    "title" : "ClaimFlow Claim API",
    "version" : "1.0.0"
  },
  "openapi" : "3.0.1",
  "paths" : {
    "/api/v1/adjusters" : {
      "get" : {
        "operationId" : "listActive",
        "parameters" : [ {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AdjusterResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter, or failed validation (see `violations`)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged, never returned)"
          }
        },
        "summary" : "List active adjusters",
        "tags" : [ "Adjusters" ]
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAdjusterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdjusterResponse"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "Location" : {
                "description" : "URL of the new resource",
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter, or failed validation (see `violations`)"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid state transition, duplicate, or a concurrent update"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged, never returned)"
          }
        },
        "summary" : "Register a claims adjuster",
        "tags" : [ "Adjusters" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "API gateway (local)",
    "url" : "http://localhost:8000"
  } ]
}